more precision in calibration
